# specs/001-price-comparison-mvp/contracts/http-api.openapi.yaml
openapi: 3.1.0
info:
  title: BestPrice MVP API
  version: 0.1.0
  description: >
    HTTP contract for the Egypt-first bilingual electronics price-comparison MVP.
    Shopper-facing endpoints exclude stale offers older than 12 hours. Admin
    endpoints are restricted to internal admins.
servers:
  - url: http://localhost:3001/api/v1
...
paths:
...
  /admin/overview:
    get:
      tags: [Admin]
      summary: Get operations summary for internal admins
      security:
        - adminSession: []
      responses:
        '200':
          description: Admin dashboard overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOverview'
        '401':
          description: Unauthorized
  /admin/sources:
    get:
      tags: [Admin]
      summary: List source freshness and crawl health
      security:
        - adminSession: []
      responses:
        '200':
          description: Source health list
          content:
            application/json:
              schema:
                type: object
                required: [sources]
                properties:
                  sources:
                    type: array
                    items:
                      $ref: '#/components/schemas/SourceHealth'
  /admin/crawl-jobs:
    post:
      tags: [Admin]
      summary: Trigger one or more crawl jobs manually
      security:
        - adminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrawlJobCreateRequest'
      responses:
        '202':
          description: Crawl jobs accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlJobCreateResponse'
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: bestprice_admin_session
  schemas:
...
    AdminOverview:
      type: object
      required:
        - sourceCount
        - staleSourceCount
        - failedJobCount
        - unmatchedProductCount
        - generatedAt
      properties:
        sourceCount:
          type: integer
        staleSourceCount:
          type: integer
        failedJobCount:
          type: integer
        unmatchedProductCount:
          type: integer
        generatedAt:
          type: string
          format: date-time
    SourceHealth:
      type: object
      required:
        - sourceAdapterId
        - store
        - status
        - lastSuccessfulCrawlAt
        - staleOfferCount
      properties:
        sourceAdapterId:
          type: string
          format: uuid
        store:
          type: string
        status:
          type: string
          enum: [healthy, warning, stale, failing]
        lastSuccessfulCrawlAt:
          type: string
          format: date-time
        staleOfferCount:
          type: integer
        latestJobStatus:
          type: string
          enum: [queued, running, succeeded, partial, failed, cancelled]
    CrawlJobCreateRequest:
      type: object
      required: [adapterIds]
      properties:
        adapterIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        reason:
          type: string
          maxLength: 200
    CrawlJobCreateResponse:
      type: object
      required: [jobIds, status]
      properties:
        jobIds:
          type: array
          items:
            type: string
            format: uuid
        status:
          type: string
          enum: [accepted]
...